Changed a number of rules in order to restrict the choice of Acc (now we have to many Acc readings).
Added "OR VFIN" to
> MAP (@DN>) TARGET (Dem Acc) IF (*1 (N Acc) BARRIER Ill OR Com OR Loc OR Ess OR VFIN);
since it failed to find topicalised objects.

Added @SUBJ-QH to the NUMERALS set, the old rule picked dates that had already been correctly identified as complem
ents to preceeding nouns, hence no candidates for heads of genitives to the right.
> SELECT Gen IF (-1 CC)(-2 Cmpnd)(*-3 (NUMERALS @SUBJ-QH) BARRIER NPNH OR PUNCT LINK 0 Nom);

Added LEFT to the barrier in the following rule:
>       (*1C V-PASS BARRIER LEFT OR V-NOT-AUX OR NP-HEAD OR Po OR CS OR CLB LINK NOT 0 Inf);

Added CC to the following barrier:
> REMOVE (Sg Nom) IF (*-1 VFIN BARRIER CC OR CS OR CP OR CONTRA OR REALCLB OR go OR ("dego")

Removed this rule, it seems:
< SELECT (@OBJ) IF (-1 ST-VERB)(0 TIME OR ROUTE)(NOT *-2C (@OBJ) BARRIER CS OR CP)(NOT *1C (@OBJ) BARRIER TV OR CP
OR CS);
<     ## Lea leamaÅ¡ liiggÃ¡s oanehis Ã¡igodat oaÅ¾Å¾ut bohtosiid oidnosii, ÄÃ¡jehuvvo
<     ## ahte easttadeaddji buohcanjÃ¡vkan dÃ¡rbbaha Ã¡iggi vai vÃ¡ikkuha nu movt lea
<     ## sÃ¡valdat.
<     ## Olbmot juige ija. # It should not take this one.

I don't quite remember why (this was collective work)


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@8269 c7155fb1-f0a7-4240-a2fc-2600b6f42f90