Added (OR VFIN) to the last constraint in this rule, so it will work even if there is no comma between the Acc and the verb. #SELECT Acc IF (NOT *-1 Pr OR Num BARRIER NPNH)(0 NOTIME OR Pron) 	(NOT 0 Nom OR Loc OR Adv OR VERB) 	(*1 CLB BARRIER NOT-ADV-PCLE LINK NOT *1 Gen BARRIER NPNH); SELECT Acc IF (NOT *-1 Pr OR Num BARRIER NPNH)(0 NOTIME OR Pron) 	(NOT 0 Nom OR Loc OR Adv OR VERB) 	(*1 CLB OR VFIN BARRIER NOT-ADV-PCLE LINK NOT *1 Gen BARRIER NPNH); 	## Jus divÅ¡Å¡ohas roasmmohuvvÃ¡ dahje oaÅ¾Å¾u duoÄalaÅ¡ vÃ¡ttisvuoÄaid,... 	## Liggen dihtii viesuideaset, ferte ollusat muoraid boaldit. 	## PolitihkalaÅ¡ sÃ¡valdat dÃ¡hkidit sÃ¡mi geavaheddjiid bÃ¡lvalusaid dÃ¡si lea maid Äuovvuluvvon ieÅ¡guÄet stuorradiggedieÄÃ¡husain.
Changed V-TRANS-ACT with V-TRANS-ACT-NOT-ACT in this rule:
REMOVE (N Gen) IF (*-1 V-TRANS-ACT-NOT-ACT OR TIME BARRIER NPNH)(NOT -1 PrfPrc LINK -1 (@OBJ))(1 Gen LINK *1 Po OR NOT-ACC-GEN BARRIER VFIN OR S-BOUNDARY);
It shouldnÂ´t take "apotehka" in this rule:
"...go doavttir lea dahkan biehttaleami apotehka vuoigatvuoÄa vuostÃ¡."


Added (LINK NOT 0 N) to this rule to avoid OPRED-VERB wich are derivated nouns like in "SÃ¡mi geavaheddjiid bÃ¡lvalusaid vuoÄÄun..":
REMOVE Gen IF (*-1 OPRED-VERB BARRIER NPNH LINK NOT 0 N LINK NOT *-1 NP-HEAD-ACC
	BARRIER V-NOT-AUX OR CC OR S-BOUNDARY)(1 Ess);
	## Don galggat gohÄodit su Jesusin.

For the same reason I changed (LINK NOT 0 Actio) with (LINK NOT 0 N) in these five rules. Analysator doesnÂ´t recognize "geavaheaddji" as Actor! It should be done something with that. Made a new set ready for it:
SET ST-VERB-NOT-ACT = ST-VERB - Actio - Actor ;

SELECT Acc IF (*-1 OPRED-VERB BARRIER NPNHA LINK NOT 0 Pass OR Actio LINK
	NOT *-1 NP-HEAD-ACC BARRIER V-NOT-AUX OR S-BOUNDARY)(NOT 0 TIME)
	(NOT *1 Acc BARRIER NPNH OR PUNCT OR CC OR ("/"));
	## DÃ¡Å¾at eai atnÃ¡n sÃ¡miid buoret go beanan.

SELECT Acc IF (*-1 V-TRANS-ACT-NOT-ACT BARRIER NOT-ADV-PCLE OR CLB LINK NOT 0 N LINK NOT *-1 NP-HEAD-ACC
	BARRIER V-NOT-AUX OR S-BOUNDARY)(1 Ess)(NOT 2 Inf);
	## Mun oidnen MÃ¡reha mÃ¡nnÃ¡n.
	## Mun oidnen MÃ¡reha boahtime.

SELECT (@OBJ) IF (*-1 ST-VERB-NOT-ACTIO BARRIER S-BOUNDARY OR (@OBJ) LINK NOT 0 N
	LINK NOT *-1 (@OBJ) BARRIER S-BOUNDARY)(NOT 0 Adv)
	(NOT 0 Nom LINK *-1 (TV N Gen) BARRIER NPNH)(NOT 0 TIME)
	(NOT 0 Dem LINK *1 N BARRIER NPNH)(NOT 0 Loc LINK *-1 (IV) BARRIER VERB)
	(NOT 0 (Acc PxSg3) LINK 0 Loc)(NOT 1 V-TRANS-ACT)
	(NOT *1 (@OBJ) BARRIER CC OR S-BOUNDARY OR COMMA)
	(NOT 0 ("nr") LINK -1 N OR A OR ACR OR ABBR);
	## Mun oidnen guovÅ¾Å¾a golbma kilomehtara dÃ¡bbelis.

SELECT (@OBJ) IF (*-1 ST-VERB-NOT-ACT BARRIER S-BOUNDARY OR VERB LINK NOT 0 N)
	(NOT *-1 (@OBJ) BARRIER S-BOUNDARY)(NOT 0 Loc)(NOT 0 Dem LINK *1 N BARRIER NPNH)
	(*1 ADVLCASE BARRIER NOT-ADV-PCLE LINK NOT *1 (@OBJ) BARRIER S-BOUNDARY);
	## De olahit olbmot luosa buoÄÄogeÅ¾iin.
In this one I also removed (OR go) from the first barrier.

REMOVE Acc IF (NOT *-1 V-TRANS-ACT-NOT-ACT BARRIER PUNCT)(NOT *-1 CS LINK *1 V-TRANS-ACT BARRIER Acc)(NOT *-1 CS BARRIER V-TRANS-ACT LINK *1 CS BARRIER VFIN)(NOT *-1 CS BARRIER V LINK 1 @CS-COMPL)
	(NOT *-1 (Pass PrfPrc) BARRIER NPNH LINK *-1 V-TRANS-ACT BARRIER NPNHA)
	(NOT -1 CC LINK -1 Acc LINK -1 ("ollu"))(NOT 0 TIME)
	(NOT 0 NUMERALS LINK *1 TIME OR MEASURE BARRIER NOT-A)
	(NOT *1 V-TRANS-ACT BARRIER S-BOUNDARY LINK 0 VFIN OR (@-FMAINV))
	(NOT *1 CLB BARRIER NOT-NP); # LINK *1 (Pron Acc) BARRIER NOT-ADV-PCLE);
	## Leago doppe kafea?
	## Amma duot lei Niillasa guollebiila?
	## LÃ¡hka divÅ¡Å¡ohasvuoigatvuoÄaid birra.
	## ...jus divÅ¡Å¡ohas oaÅ¾Å¾u Ã¡vkki dearvvaÅ¡vuoÄaveahkis.
	## SÃ¡medikki Äoahkkimat dollojuvvojit rabas uvssaid.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@8621 c7155fb1-f0a7-4240-a2fc-2600b6f42f90