Added PrsPrc to the set: SET V-TRANS-ACT-NOT-ACT = TV - V-PASS - Actio - Actor - PrsPrc ;
The sme-lex is now chaged so the analysator marks also Actor from odd syllable verbs. Removed (LINK NOT 0 N) from the first part of this rule, because itÂ´s not necessary any more, and because it didnÂ´t functin as it was ment (NOT ... NOT).

REMOVE Acc IF (NOT *-1 V-TRANS-ACT-NOT-ACT BARRIER PUNCT)(NOT *-1 CS LINK *1 V-TRANS-ACT BARRIER Acc)(NOT *-1 CS BARRIER V-TRANS-ACT LINK *1 CS BARRIER VFIN)(NOT *-1 CS BARRIER V LINK 1 @CS-COMPL) (NOT *-1 (Pass PrfPrc) BARRIER NPNH LINK *-1 V-TRANS-ACT BARRIER NPNHA) (NOT -1 CC LINK -1 Acc LINK -1 ("ollu"))(NOT 0 TIME) (NOT 0 NUMERALS LINK *1 TIME OR MEASURE BARRIER NOT-A) (NOT *1 V-TRANS-ACT BARRIER S-BOUNDARY LINK 0 VFIN OR (@-FMAINV)) (NOT *1 CLB BARRIER NOT-NP); # LINK *1 (Pron Acc) BARRIER NOT-ADV-PCLE);

In this sentence the problem is to find that "loktema" is Acc, the others are Gen.
"RÃ¡ÄÄehus deattuha sÃ¡mi kulturÃ¡ddejumi gelbbolaÅ¡vuoÄa loktema sÃ¡miid buorrin."

Added (OR Acc) to the last barrier of this rule, so it wonÂ´t hit "sÃ¡mi":
REMOVE (N Gen) IF (*-1 V-TRANS-ACT-NOT-ACT OR TIME BARRIER NPNH)(NOT -1 PrfPrc LINK -1 (@OBJ))(1 Gen LINK *1 Po OR NOT-ACC-GEN BARRIER Acc OR VFIN OR S-BOUNDARY);

Made a new set: NATION = "Norga" "SÃ¡pmi" (more to come).  Added ("sÃ¡pmi" N Pl) to HUMAN-GROUP.

Made a new sub.chapter, and the first rule to it:
# Accusative in a Acc/Gen string
# ------------------------------

SELECT Acc IF (0 Gen LINK NOT 0 NATION OR POLITICAL-PLACE)(1 HUMAN-GROUP LINK 0 Gen);
	## RÃ¡ÄÄehus deattuha sÃ¡mi kulturÃ¡ddejumi gelbbolaÅ¡vuoÄa loktema sÃ¡miid buorrin.

Made two new rules which I put into cycle 4, when the analysator probably has found the Acc.

SELECT Gen IF (0 Acc)(NOT 0 TIME OR ROUTE)(1C Gen)(2C Acc);
SELECT Gen IF (0 Acc)(NOT 0 TIME OR ROUTE)(1C Gen)(2C Gen)(3C Acc);


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@8670 c7155fb1-f0a7-4240-a2fc-2600b6f42f90