Wrote with Linda some comments to the verb-sets.
Removed the Comitative idiomatic expressions so they come before # Locative and comitative - Disambiguation based on verbs.

Added ("veardidit") to this rule:
SELECT Gen IF (*-1 ("vrd") OR ("veardidit") BARRIER NOT-AA)(1 Com);

Added * and "veardidit" to this rule.
#SELECT (Sg Com) IF (-1 ("vrd") OR ("veardidit"))(0 (Pl Loc));
SELECT (Sg Com) IF (*-1 ("vrd") OR ("veardidit") BARRIER Com)(0 (Pl Loc));
  	## SÃ¡mediggi lea vÃ¡ldÃ¡n badjelasas vÃ¡ldoovddasvÃ¡stÃ¡dusa sihkkarastimis doaibmadoarjaga NuortasÃ¡mi museai (vrd. Äoahkkimiin NjÃ¡vdÃ¡mis 05.06.01).
    ## Sii Äuvvot sÃ¡milÃ¡ga mearrÃ¡dusaid, veardit siskkÃ¡ldas goziheami geatnegasvuoÄain.

Added "gulahallat" and "nannet" to COM-VERB.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@8758 c7155fb1-f0a7-4240-a2fc-2600b6f42f90