Wrote with Linda some comments to the verb-sets.
Removed the Comitative idiomatic expressions so they come before # Locative and comitative - Disambiguation based on verbs.

Added ("veardidit") to this rule:
SELECT Gen IF (*-1 ("vrd") OR ("veardidit") BARRIER NOT-AA)(1 Com);

Added * and "veardidit" to this rule.
#SELECT (Sg Com) IF (-1 ("vrd") OR ("veardidit"))(0 (Pl Loc));
SELECT (Sg Com) IF (*-1 ("vrd") OR ("veardidit") BARRIER Com)(0 (Pl Loc));
  	## Sámediggi lea váldán badjelasas váldoovddasvástádusa sihkkarastimis doaibmadoarjaga Nuortasámi museai (vrd. čoahkkimiin Njávdámis 05.06.01).
    ## Sii čuvvot sámilága mearrádusaid, veardit siskkáldas goziheami geatnegasvuoÄain.

Added "gulahallat" and "nannet" to COM-VERB.