For practical and consistency reasons I changed VERB into V in most verb sets:
Changed ST-VERB to STV
Changed ST-VERB-NOT-ACT to STV-NOT-ACT
Changed LOC-VERB to LOCV
Changed COM-VERB to COMV
Changed ILL-VERB to ILLV
Changed TRANSL-VERB to TRANSLV
Changed ACC-LOC-VERB to ACC-LOCV
Changed ACC-ILL-VERB to ACC-ILLV
Changed INF-VERB to INFV
Changed ACC-INF-VERB to ACC-INFV
Changed MIELDE-VERB to MIELDE-V
Changed OPRED-VERB to OPREDV
Changed TOGETHER-VERB to TOGETHER-V
Changed SOURCE-VERB to SOURCE-V

The valency verb sets have the "V" attached to their type-of-valency symbol analogous to IV and TV such as in COMV and INFV.

Verbsets with a lexical item or a semantic classification, such as
MIELDE-V or PLACE-V retain their hyphen.
The distinction is made for practical reasons keeping in mind the legibility of set names.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@8762 c7155fb1-f0a7-4240-a2fc-2600b6f42f90