Invalidated the following rule:
#SELECT (Pl Loc) IF (0 (Sg Com));

to be able to search for further ambiguity and tricky cases (it will be useful at a later point of time though)

and put the following rules into use again:

SELECT (Pl Loc) IF (*-1 DYNAMIC-V BARRIER S-BOUNDARY)(0 (Sg Com) LINK NOT 0 TOOLS OR ABSTRACT-TOOLS OR WRITING-TOOLS OR CONCEPT OR HUMAN);
SELECT (Pl Loc) IF (0 (Sg Com) LINK NOT 0 TOOLS OR ABSTRACT-TOOLS OR WRITING-TOOLS OR CONCEPT OR HUMAN)(*1 DYNAMIC-V BARRIER S-BOUNDARY);

SELECT (Pl Loc) IF (*-1 HUMAN-SOURCE-V BARRIER S-BOUNDARY)(0 (Sg Com) LINK NOT 0 TOOLS OR ABSTRACT-TOOLS OR WRITING-TOOLS OR CONCEPT);
SELECT (Pl Loc) IF (0 (Sg Com) LINK NOT 0 TOOLS OR ABSTRACT-TOOLS OR WRITING-TOOLS OR CONCEPT)(*1 HUMAN-SOURCE-V BARRIER S-BOUNDARY);



Added the preceding V as a NOT-constraint to the following rule:

REMOVE Adv IF (NOT -1 ("seamma") OR ("seammÃ¡") OR ("mii" Pron))(NOT -1 V)(0 ("ollu"))(NOT 1 V);
  # This is a kill-all-rule...

For ollu being an Adverb, it can be preceeded by a  verb as well.
The problem is:
in front of a potential loc ollu can be Pron Indef
as a member of HEADSOFPART being the head in a partitive construction.
we have to deal with the disambiguation of com/loc first in order to deceide if we have a Pron Indef and therefore a partitive construction, or rather an Adv in   a construction where the verb is connected to the noun through lexical valency.
This is a semantic decision.

we need therfore a rule later:

SELECT Adv IF (-1 V)(0 ("ollu"))(*1C Com BARRIER NPNH);
    ## Lean vuodjÃ¡n ollu biillain.
    ## Lean borran ollu gÃ¡ffaliin.

REMOVE Adv IF (-1 V)(0 ("ollu"))(*1C Loc BARRIER NPNH);
    # The actual HEADOFPARTS ollu..
    ## Lean borran ollu gÃ¡hkkuin.

But still comitative and locative have to be disambiguated before:

That means, the head of part rules should be moved, or at least "ollu"


Did not move the following rules because in their position in front of the accusative, they seemed to have an important funcion:

SELECT Loc IF (*-1 HEADOFPARTS BARRIER NPNH OR COMMA);
	## GuhtemuÅ¡ dainna bivdovugiin guhkimus leaÅ¡ gevahuvvon...
	## Dat sÃ¡httet leat oassin Ã¡rvvoÅ¡tallamiin. #If this does not work: maybe: REMOVE Acc?

SELECT Loc IF (*-1 HEADOFPARTS LINK 1 Gen BARRIER NPNH OR COMMA);
  # Just in case there is a genitive which still is an accusative as well and therefore NPNH and ruled out by the rule before.

SELECT Loc IF (-3 HEADOFPARTS LINK 1 Ill LINK 1 (PrfPrc @PrcN>));
SELECT Loc IF (-3 HEADOFPARTS LINK 1 Com LINK 1 (PrfPrc @PrcN>));
SELECT Loc IF (-3 HEADOFPARTS LINK 1 Loc LINK 1 (PrfPrc @PrcN>));

SELECT Loc IF (*-1 HEADOFPARTS BARRIER NPNH OR COMMA);
	## GuhtemuÅ¡ dainna bivdovugiin guhkimus leaÅ¡ gevahuvvon...
	## Dat sÃ¡httet leat oassin Ã¡rvvoÅ¡tallamiin.


SELECT Loc IF (*-1 HEADOFPARTS BARRIER NPNH OR COMMA)(0 Pl or Pl3)(NOT 0 Num OR Ord) ;
    	## Mun dovdan vihtta proseantta dainna.
	## Dat maid gohÄodit normÃ¡la giellaovdÃ¡neapmi dÃ¡hpÃ¡huvvÃ¡ mÃ¡nÃ¡in geat
	## bajÃ¡sÅ¡addet eanet go ovttain gielain go seamma jahkÃ¡saÅ¡ ovttagielat mÃ¡nÃ¡in.
  # This rule should be tested. In: zcorp/dis/sme/all/1.429.txt (8272).



Instead, I removed "ollu" and "ollut" from the HEADOFPARTS set, and made particular rules for it, and placed the in the late locative section:

SELECT Loc IF (*-1 ("ollu") BARRIER NPNH OR COMMA);
	## Ollu dainna bivdovugiin guhkimus leaÅ¡ gevahuvvon...

SELECT Loc IF (*-1 ("ollu") LINK 1 Gen BARRIER NPNH OR COMMA);
  # Just in case there is a genitive which still is an accusative as well and therefore NPNH and ruled out by the rule before.

SELECT Loc IF (-3 ("ollu") LINK 1 Ill LINK 1 (PrfPrc @PrcN>));
SELECT Loc IF (-3 ("ollu") LINK 1 Com LINK 1 (PrfPrc @PrcN>));
SELECT Loc IF (-3 ("ollu") LINK 1 Loc LINK 1 (PrfPrc @PrcN>));


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@8798 c7155fb1-f0a7-4240-a2fc-2600b6f42f90