Added items to ORGANIZATION, ABSTRACT-TEXT
Changed the following rule, it hit where it should not have. It selected Pl Loc because of a preceeding DYNAMIC-V even though there was a CS between the verb and its assumed argument(which it was not). The BARRIER did not work as it was supposed to in this place.

<SELECT (Pl Loc) IF (*-1 DYNAMIC-V BARRIER SV-BOUNDARY)(0 (Sg Com) BARRIER CS LINK NOT 0 TOOLS OR ABSTRACT-TOOLS OR WRITING-TOOLS OR CONCEPT OR HUMAN);

>SELECT (Pl Loc) IF (0 (Sg Com) BARRIER CS LINK NOT 0 TOOLS OR ABSTRACT-TOOLS OR WRITING-TOOLS OR CONCEPT OR HUMAN LINK *-1 DYNAMIC-V BARRIER SV-BOUNDARY);

Did the same thing with the following rule:

<SELECT (Pl Loc) IF (*-1 HUMAN-SOURCE-V BARRIER SV-BOUNDARY)(0 (Sg Com) LINK NOT 0 TOOLS OR ABSTRACT-TOOLS OR WRITING-TOOLS OR CONCEPT);

>SELECT (Pl Loc) IF (0 (Sg Com) LINK NOT 0 TOOLS OR ABSTRACT-TOOLS OR WRITING-TOOLS OR CONCEPT LINK *-1 HUMAN-SOURCE-V BARRIER SV-BOUNDARY);


IMPORTANT:
==========

Those rules need to be checked in concrete example sentences, especially complex ones. They do not take into account more complex constructions. Those rules simply take any NP and do not have NPs as border. But: NPs can themselves be predicates for NPs and select a particular case.
Idiomatic expressions are not taken into account either.
Example "váldit vuhtii". In this case it still belongs to the DYNAMIC-V set, but there might be other cases where it doesnÂ´t even though the verb without its idiomatic particle (adverb) does.