Tried to solve the ComLoc with Actio Pl. Made a set with most actios: LIST EVENT-TOOL-ACTIO = (Actio) + lexicalized actios. These can be both tool and event. Actios which cannot be both, we put in ABSTRACT-TOOLS, TOOLS or EVENT, like: eallin (TOOLS), boazodoallo#Å¡iehtadallan (EVENT) jnv.
Added the set to the rules for place-verbs and dynamic-verbs. It seems to work, like in these sentences, most of them from the bible: Mun lean doarrádallan jápmimiin ja goddimiin sin. Dan son celkkii almmuhan dihtii makkár jápmimiin Biehtár galggai gudnejahttit Ipmila. Son duoÄaÅ¡tii dan addimiin sidjiide Bassi VuoiÅÅa. Mun vulgen viehkamiin.

Some verbs donÂ´t fit into the system: "aassálastit" and "eallit" are ABSTR-ACTIVITY, but it can take Actio in Loc: "oassálastit viehkamiin". Same situation with some of the ACTIVITY-V: dubmet, juhkat, veahkehit. So the Actio-problem is not quit solved, we have to modify it.

I have noticed a problem about the verb-sets. When a verb gets Der, then it get analized as the basic verb, even if it it lexicalized. F.i. "čuovvulit" get analized as "čuovvut" which is a movement-verb, "čuovvulit" is not.

Made a new verbset: SET FMAINV = @-FMAINV | @+FMAINV ;
which can be useful in many rules to localize the main-verb. Changed a couple of rules due to this, and also the SV-BOUNDARY. V-MAIN only tell that the verb cannot be an Aux or Neg or "orrut".

The analysator doesnÂ´t always give the FMAINV analyse when it should, like in: In máhte sámi.

Some of the rules based upon verb-sets, have constraint with subject link verb-set, and they can hit actio instead of the mainverb if the actio-word isnÂ´t lexicalized as noun. The place-verb-rule hits f.i. "Mun vuiten viehkamiin". To prevent it, and also hitting PrsPrc used as adjektiv, I added (LINK 0 FMAIN LINK NOT 0 Actio) to many verb-rules.

Removed this rule, it hits to widely, and should not hit: Universitehta lea njunnoÅ¡is sámi áÅ¡Å¡iid dutkamiin.
#SELECT (Sg Com) IF (*-1 (N Sg Loc) BARRIER NPNH OR PUNCT-LEFT OR PUNCT-RIGHT OR CLB OR COMMA LINK *-1 IV BARRIER TV)(0 N);
  	## Várrepresideanta lea maid 05.01.00 leamaÅ¡ čoahkkimis Sosiála- ja dearvvasvuoÄadepartemeanttain čilgen dihte Sámedikki juolludemiid sámi dearvvasvuoÄa- ja sosiálaplána čuovvuleamis .
  	## Soai leigga čohkkámin fatnasis áhčiineaskka ja čiktimin firpmiid.

Removed this rule:
#SELECT (Pl Loc) IF (*-1 CC OR ("/") BARRIER NPNH LINK *-1 Loc LINK NOT 0 Com BARRIER NPNH);
	## Seamma olmmoÅ¡ áimmuhuÅ¡Å¡á máÅga dáin doaimmain dahje rollain.
	## Sámi oahpahussuorgi stivrejuvvo daid rámmaid mielde maid lágat ja njuolggadusat bidjet, ja daid ulbmiliid ja viÅ¡uvnnaid mielde mat ovdanbohtet Sámediggeplánas ja eará oahpahus- ja oahppoplánain ja strategiijain.
	## Mii maid muÄui fuomáÅ¡uhttit ahte ollu lohkoávdnasiin mat fátmmastit dahje čilgejit sámediggeválgga , dain lea sáhka smávva loguin / vuollegis lohkoárvvuin.