Tried to solve the ComLoc with Actio Pl. Made a set with most actios: LIST EVENT-TOOL-ACTIO = (Actio) + lexicalized actios. These can be both tool and event. Actios which cannot be both, we put in ABSTRACT-TOOLS, TOOLS or EVENT, like: eallin (TOOLS), boazodoallo#Å¡iehtadallan (EVENT) jnv.
Added the set to the rules for place-verbs and dynamic-verbs. It seems to work, like in these sentences, most of them from the bible: Mun lean doarrÃ¡dallan jÃ¡pmimiin ja goddimiin sin. Dan son celkkii almmuhan dihtii makkÃ¡r jÃ¡pmimiin BiehtÃ¡r galggai gudnejahttit Ipmila. Son duoÄaÅ¡tii dan addimiin sidjiide Bassi VuoiÅÅa. Mun vulgen viehkamiin.

Some verbs donÂ´t fit into the system: "aassÃ¡lastit" and "eallit" are ABSTR-ACTIVITY, but it can take Actio in Loc: "oassÃ¡lastit viehkamiin". Same situation with some of the ACTIVITY-V: dubmet, juhkat, veahkehit. So the Actio-problem is not quit solved, we have to modify it.

I have noticed a problem about the verb-sets. When a verb gets Der, then it get analized as the basic verb, even if it it lexicalized. F.i. "Äuovvulit" get analized as "Äuovvut" which is a movement-verb, "Äuovvulit" is not.

Made a new verbset: SET FMAINV = @-FMAINV | @+FMAINV ;
which can be useful in many rules to localize the main-verb. Changed a couple of rules due to this, and also the SV-BOUNDARY. V-MAIN only tell that the verb cannot be an Aux or Neg or "orrut".

The analysator doesnÂ´t always give the FMAINV analyse when it should, like in: In mÃ¡hte sÃ¡mi.

Some of the rules based upon verb-sets, have constraint with subject link verb-set, and they can hit actio instead of the mainverb if the actio-word isnÂ´t lexicalized as noun. The place-verb-rule hits f.i. "Mun vuiten viehkamiin". To prevent it, and also hitting PrsPrc used as adjektiv, I added (LINK 0 FMAIN LINK NOT 0 Actio) to many verb-rules.

Removed this rule, it hits to widely, and should not hit: Universitehta lea njunnoÅ¡is sÃ¡mi Ã¡Å¡Å¡iid dutkamiin.
#SELECT (Sg Com) IF (*-1 (N Sg Loc) BARRIER NPNH OR PUNCT-LEFT OR PUNCT-RIGHT OR CLB OR COMMA LINK *-1 IV BARRIER TV)(0 N);
  	## VÃ¡rrepresideanta lea maid 05.01.00 leamaÅ¡ Äoahkkimis SosiÃ¡la- ja dearvvasvuoÄadepartemeanttain Äilgen dihte SÃ¡medikki juolludemiid sÃ¡mi dearvvasvuoÄa- ja sosiÃ¡laplÃ¡na Äuovvuleamis .
  	## Soai leigga ÄohkkÃ¡min fatnasis Ã¡hÄiineaskka ja Äiktimin firpmiid.

Removed this rule:
#SELECT (Pl Loc) IF (*-1 CC OR ("/") BARRIER NPNH LINK *-1 Loc LINK NOT 0 Com BARRIER NPNH);
	## Seamma olmmoÅ¡ Ã¡immuhuÅ¡Å¡Ã¡ mÃ¡Åga dÃ¡in doaimmain dahje rollain.
	## SÃ¡mi oahpahussuorgi stivrejuvvo daid rÃ¡mmaid mielde maid lÃ¡gat ja njuolggadusat bidjet, ja daid ulbmiliid ja viÅ¡uvnnaid mielde mat ovdanbohtet SÃ¡mediggeplÃ¡nas ja earÃ¡ oahpahus- ja oahppoplÃ¡nain ja strategiijain.
	## Mii maid muÄui fuomÃ¡Å¡uhttit ahte ollu lohkoÃ¡vdnasiin mat fÃ¡tmmastit dahje Äilgejit sÃ¡mediggevÃ¡lgga , dain lea sÃ¡hka smÃ¡vva loguin / vuollegis lohkoÃ¡rvvuin.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@8872 c7155fb1-f0a7-4240-a2fc-2600b6f42f90