Added ("/") to the set of coordinators since the following rule had been removed before:
#SELECT (Pl Loc) IF (*-1 CC OR ("/") BARRIER NPNH LINK *-1 Loc LINK NOT 0 Com BARRIER NPNH);


The following rule does not pay attention to coordination, it simply hits on the coordinated item as well,

<SELECT (Pl Loc) IF (*-1 LOC-COMPL-NOUN BARRIER SV-BOUNDARY LINK NOT 0 ("ovda#mearka" Ess))(0 (Sg Com));
>SELECT (Pl Loc) IF (*-1 LOC-COMPL-NOUN BARRIER SV-BOUNDARY OR CRD LINK NOT 0 ("ovda#mearka" Ess))(0 (Sg Com));

which in some cases might be nice such as in:
## Mii maid muÄui fuomáÅ¡uhttit ahte ollu lohkoávdnasiin mat fátmmastit dahje čilgejit sámediggeválgga , dain lea sáhka smávva loguin / vuollegis lohkoárvvuin

The problem becomes clear though when we extend the coordination from NP to VP coordination, where not loguin and lohkoárvvuin are coordinated, but the "/" introduces a new sentence, and the case of "lohkoárvvuin" does definitely not depend on sáhka as its predicate (LOC-COMPL-NOUN):

## Mii maid muÄui fuomáÅ¡uhttit ahte ollu lohkoávdnasiin mat fátmmastit dahje čilgejit sámediggeválgga , dain lea sáhka smávva loguin / vuollegis lohkoárvvuin leat buorre boaÄus.

Added CRD to those rules as well:

SELECT (Pl Loc) IF (*-1 DYNAMIC-V BARRIER SV-BOUNDARY OR CRD)(0 (Sg Com) LINK NOT 0 TOOLS OR ABSTR-TOOLS OR WRITING-TOOLS OR CONCEPT OR HUMAN);
SELECT (Pl Loc) IF (0 (Sg Com) LINK NOT 0 TOOLS OR ABSTR-TOOLS OR WRITING-TOOLS OR CONCEPT OR HUMAN)(*1 DYNAMIC-V BARRIER SV-BOUNDARY OR CRD LINK 0 FMAINV);

The point is: as long as coordinators cannot be identified clearly as coordinators between noun phrases, predicate/valency rules and coordination rules need to be seperate.

I added the following constraint to ensure that:

(NOT *1 VFIN OR AUX BARRIER SV-BOUNDARY)

Changed the following rules:

<SELECT Com IF (*-1 CC OR COMMA BARRIER NPNH LINK *-1C Com BARRIER WORD);
>SELECT Com IF (*-1 CRD BARRIER NPNH LINK *-1C Com BARRIER WORD)(NOT *1 VFIN OR AUX BARRIER SV-BOUNDARY);

<SELECT (Sg Loc) IF (1 CC OR ("/"))(*2C Loc BARRIER NPNH);
>SELECT (Sg Loc) IF (1 CRD)(*2C Loc BARRIER NPNH)(NOT *1 VFIN OR AUX BARRIER SV-BOUNDARY);

<SELECT (Sg Loc) IF (*-1 CC OR ("/") BARRIER NPNH LINK -1C Loc);
>SELECT (Sg Loc) IF (*-1 CRD BARRIER NPNH LINK -1C Loc)(NOT *1 VFIN OR AUX BARRIER SV-BOUNDARY);

<SELECT (Pl Loc) IF (1 CC OR (Neg Foc) OR ("/"))(*2 Loc BARRIER NPNH LINK NOT 0 Com);
>SELECT (Pl Loc) IF (1 CRD)(*2 Loc BARRIER NPNH LINK NOT 0 Com)(NOT *1 VFIN OR AUX BARRIER SV-BOUNDARY);

<SELECT Loc IF (*-1C Loc LINK 1 COMMA OR CC OR ("/"))(0 (Sg Com) BARRIER NPNH)(1 COMMA OR CC OR (Neg Foc) LINK *1C Loc BARRIER NPNH);
>SELECT Loc IF (*-1C Loc LINK 1 CRD)(0 (Sg Com) BARRIER NPNH)(1 COMMA OR CC OR (Neg Foc) LINK *1C Loc BARRIER NPNH)(NOT *1 VFIN OR AUX BARRIER SV-BOUNDARY);

<SELECT Loc IF (*-1C Loc LINK *1 Rel BARRIER Com OR S-BOUNDARY)(0 Com LINK *-1 CC OR COMMA OR (Neg Foc) OR ("/") BARRIER NPNH)(NOT *1 VFIN OR AUX BARRIER SV-BOUNDARY);
>SELECT Loc IF (*-1C Loc LINK *1 Rel BARRIER Com OR S-BOUNDARY)(0 Com LINK *-1 CRD BARRIER NPNH)(NOT *1 VFIN OR AUX BARRIER SV-BOUNDARY);

and some others

Put the following rule:

SELECT (Pl Loc) IF (*-1 CRD BARRIER NPNH LINK *-1 Loc LINK NOT 0 Com BARRIER NPNH)(NOT *1 VFIN OR AUX BARRIER SV-BOUNDARY);

into use again since it is the one choosing Loc in the case of that example sentence.