Worked on Range:
Went through the numeral mappings which were available for Range and changed them if necessary:

Changed this one to allow constructions such as:
## Geassemánu 8.b. 2001-2002 olbmo ledje...
where Range is not a date apposition

<MAP (@APP) TARGET Num IF (-3 MANNU)(-2 Ord)(-1 ("beaivi") OR ("b" ABBR));
>MAP (@APP) TARGET Num IF (-3 MANNU)(-2 Ord)(-1 ("beaivi") OR ("b" ABBR))(NOT 0 Range);
  	## Geassemánu 8.b. 2001 láhkaásahus nr 676 oktagaslaÅ¡ plánaid birra dearvvaÅ¡vuoÄalágaid vuoÄul lea addojuvvon láhkamearrádusaid láhkavuoÄuin.

did not change this one:
MAP (@NumQ<) TARGET (Num) IF (-1 Num)(0 NUMUNIT);
  	## Dat lea 4 miljovnna deajakohpa.

Left those ones as they are, even though we could get problems with sentence final year-timeadverbials:

MAP (@NumN<) TARGET Num IF (-1 N OR ACR LINK NOT 0 (@NQ<))(0 CARDINALS)(1 EOS)(NOT -1 NUMMAR OR MANNU);
  	## Dat lea dieÄáhus 4.
  	## Son dagai 1000 diimmu 2005:s.
	## Viidváseappot mearrida dearvvaÅ¡vuoÄabargiidlága Â§ 10 vuosttas laÄÄasa...
 # Should hit numerals that only have to do with numbering, not with amount.
MAP (@NumN<) TARGET Num IF (-2 N)(-1 NUMMAR OR MANNU)(0 CARDINALS)(1 EOS);
  	## Dat lea dieÄáhus nr. 4.

Changed the example for that rule and made the mapping available only for numerals not range numerals:

<MAP (@NumN<) TARGET Num IF (-1 N OR A OR ACR LINK NOT 0 (@NQ<))(0 CARDINALS)(NOT 0 ("oktii"))(NOT 1 NPROP OR A OR ACR)(NOT -1 NUMMAR OR MANNU);
>MAP (@NumN<) TARGET (Num Nom) IF (-1 N OR A OR ACR LINK NOT 0 (@NQ<))(0 CARDINALS)(NOT 0 ("oktii"))(NOT 1 NPROP OR A OR ACR)(NOT -1 NUMMAR OR MANNU);
  	## Girji 4 lei hui buorre girji.
  #This one is still too strict. It's getting better..

<MAP (@NumN<) TARGET Num IF (-1 N OR A OR ACR LINK NOT 0 (@NQ<))(0 CARDINALS LINK NOT 0 ADVLCASE)(NOT 0 ("oktii") OR ("okta") OR ("1"))(1 Nom)(NOT -1 NUMMAR OR MANNU);
>MAP (@NumN<) TARGET (Num Nom) IF (-1 N OR A OR ACR LINK NOT 0 (@NQ<))(0 CARDINALS LINK NOT 0 ADVLCASE)(NOT 0 ("oktii") OR ("okta") OR ("1"))(1 Nom)(NOT -1 NUMMAR OR MANNU);

Tested this one, we do not want the Range tag here either:

<MAP (@NumN<) TARGET Num IF (0 CARDINALS) (-1 ("s") OR ("siidu") OR ("Â§") OR ("paragráfa") OR ("kapihttal") OR ("kapihtal") OR HEADOFPARTS OR ("jahki") OR GEASSET LINK NOT 0 (@NQ<));
	## Dat lea s. 240
	## Dat leat s. 240-250. # evtl. LAHKA
>MAP (@NumN<) TARGET (Num Nom) IF (0 CARDINALS) (-1 ("s") OR ("siidu") OR ("Â§") OR ("paragráfa") OR ("kapihttal") OR ("kapihtal") OR HEADOFPARTS OR ("jahki") OR GEASSET LINK NOT 0 (@NQ<));
<MAP (@NumN<) TARGET Num IF (-3 ("s") OR ("siidu") OR ("Â§") OR ("paragráfa") OR ("kapihttal") OR HEADOFPARTS OR GEASSET LINK NOT 0 (@NQ<)) (-2 Num)(-1 CC)(0 CARDINALS);
>MAP (@NumN<) TARGET (Num Nom) IF (-3 ("s") OR ("siidu") OR ("Â§") OR ("paragráfa") OR ("kapihttal") OR HEADOFPARTS OR GEASSET LINK NOT 0 (@NQ<)) (-2 Num)(-1 CC)(0 CARDINALS);
	## ...gč. dearvvaÅ¡vuoÄabargiidlága kapihttal 6 ja 7.

<MAP (@QN>) TARGET Num IF (*1 Ess BARRIER NPNH LINK NOT 0 (V Actio));
>MAP (@QN>) TARGET (Num Nom) IF (*1 Ess BARRIER NPNH LINK NOT 0 (V Actio));
	##

<MAP (@QN<) TARGET Num IF (-1 ("diibmu" Nom) OR ("dbm" Nom));
>MAP (@QN<) TARGET (Num Nom) IF (-1 ("diibmu" Nom) OR ("dbm" Nom));
	## Mun boaÄán diibmu vihtta/viÄas.
	## Mun boaÄán diibmu 5-6.

<MAP (@QN>) TARGET Num IF (1 ("jahkásaÅ¡"));
>MAP (@QN>) TARGET (Num Nom) IF (1 ("jahkásaÅ¡"));
	## Golmma-njealje jahkásaÅ¡ mánná biddjojuvvo giissá sisa čohkkut.

Added the following rule:

MAP (@ADVL) TARGET Range ;

in order to give Range some useful meaning and allow for an adverbial mapping.
In the following sentence Range is a time adverbial:
## MáÅggabealatvuohta ja ovttadássásaÅ¡vuohta ráÄÄehusa dearvvaÅ¡vuoÄa- ja sosiálabálvalusaid doaibmaplána Norgga sápmelaččaid várás 2002-2005.
We should deceide which kind of syntactic dependency we want.
With Loc/Com we simply say they are @ADVL even if they postmodify a noun.
What about this range-numeral? It postmodifies the noun, but localy seen is a time adverbial. Here we get into a argument-adjunct discussion.
Right now we still get both:


"<2002-2005>" S:8076, 8076, 12579
        "2002-2005" Num Nom S:2064 @NumN<
        "2002-2005" Num Range S:2128 @ADVL


The Oslo-Bergen tagger does not have a solution for that. It simply assigns the quantifyier tag.


Added the following rule:

SELECT ("stuorradigge#dieÄáhus") IF (0  ("stuorra#digge#dieÄáhus"));