Working with Acc(Gen. Added some words to sets.
Made a new rule getting rid of Pass:
>REMOVE (V TV Pass) IF (0 PrfPrc)(1 Acc)(NOT 2 N);
    ## Dan olis leat dorjon oahppodoaimmaid mat ovddidit sámi kulturáddejumi.

Had to change Pass with (V TV Pass) in som verbsets to get them to function!

Made a new rule:
>REMOVE Nom IF (-1 Gen)(NOT 0 ABBR)(*1 (N Nom) BARRIER NP-BOUNDARY OR VERB OR Adv OR CC OR ADVLCASE OR PUNCT OR CLB LINK NOT 1 N);
    ## Romssa universitehta Sámi dearvvaÅ¡vuoÄadutkama guovddáÅ¡ (sajuÅ¡tuvvo KáráÅ¡johkii) ásahuvvo jagis 2001 sosiála- ja dearvvaÅ¡vuoÄadepartemeantta ruhtademiin.

Added (LINK NOT 0 NATION) to the Prop because that sámi departments often are written as Prop (Universitehta Sámi guovddáÅ¡):
>REMOVE Gen IF (0 (N Acc) LINK NOT 0 NATION)(1 Prop LINK NOT 0 NATION);

Of the same reason I added (LINK NOT 0 NATION OR GENERAL-PLACE OR ORGANIZATION OR EDUCATION) to this:
>REMOVE Gen IF (0 Nom LINK NOT 0 NATION OR GENERAL-PLACE OR ORGANIZATION OR EDUCATION)(1 (Prop Gen))(*2 N OR Po BARRIER NPNH);
	## RáÄÄehusa dearvvaÅ¡vuoÄa- ja sosiálabálvalusaid doaibmaplána Norgga sápmelaččaid várás .

Also when "sámi" isnÂ´t a Prop, it can be Gen in front of it. Added (NOT 0 OR ORGANIZATION OR EDUCATION) to htis rule:
>REMOVE Gen IF (0 Acc LINK NOT 0 NATION OR GENERAL-PLACE OR ORGANIZATION OR EDUCATION)(1 NATION OR POLITICAL-PLACE);