Added the following rule to:
REMOVE Ill IF (0 Adv LINK 0C ("lÃ¡hki") OR ("lÃ¡hkai");
    ## MÃ¡Ågii vurdojuvvo, vaikko vel ii jitnosit daddjoge, ahte almmolaÅ¡ bÃ¡lvalusaid geavaheaddjit livÄÄe seammÃ¡lÃ¡gant maiddÃ¡i kultuvrra dÃ¡fus dahje sii unnimusat lÃ¡htteÅ¡edje dÃ¡Å¾aid vÃ¡ldooasi lÃ¡hkai, go mun barggan dÃ¡kkÃ¡r bÃ¡lvalusaid.

Added the following rule to make the last of the sentences work:

MAP (@+FAUXV @+FMAINV) TARGET VFIN IF (NOT *-1 VFIN BARRIER S-BOUNDARY) (0 AUX-OR-MAIN  OR MOD-ASP);
	## Maid MÃ¡ret Ã¡iggui goarrumis?
	## Mun galggan ruoktot.


There was a problem:
"galget" did not receive a syntactic tag in a very clear position in the sentence:

BÃ¡ikkÃ¡laÅ¡ bargit ja hÃ¡lddahus galget maid sÃ¡httit oahpÃ¡smuvvat persovnnalaÅ¡ fierpmÃ¡dagaiguin, mat doibmet sÃ¡miid veahkkevuogÃ¡dahkan.

The reason was that a very general rule had been removed on 07.11 (version 1.864):

MAP (@+FAUXV) TARGET VFIN IF (0 AUX);
	## Don Å¡attat vel joksat SirpmÃ¡s gohcciid.

And replaced by the following rule:

MAP (@+FAUXV) TARGET AUX IF (NOT *-1 AUX BARRIER NOT-AA)(0 VFIN)(*1 Inf OR ConNeg OR PrfPrc BARRIER S-BOUNDARY OR VFIN);
    ## Don Å¡attat vel joksat SirpmÃ¡s gohcciid.
	## In sÃ¡hte olles gÃ¡hku borrat.


"maid" was causing a problem in such an early rule, since it can be both be an adverb and a pronoun. Used in a barrier, it causes a problem. Probably, we will get the same problem with other adverbs that can be other POS. We will have to add them to the same set and disambiguate later, or design rules only for those and disambiguate later.

That is how the set looks like now:


SET NOT-AA = WORD - A - Adv | ("maid") ;


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@9632 c7155fb1-f0a7-4240-a2fc-2600b6f42f90