Added the following rule to:
REMOVE Ill IF (0 Adv LINK 0C ("láhki") OR ("láhkai");
    ## MáÅgii vurdojuvvo, vaikko vel ii jitnosit daddjoge, ahte almmolaÅ¡ bálvalusaid geavaheaddjit livčče seammálágant maiddái kultuvrra dáfus dahje sii unnimusat láhtteÅ¡edje dážaid váldooasi láhkai, go mun barggan dákkár bálvalusaid.

Added the following rule to make the last of the sentences work:

MAP (@+FAUXV @+FMAINV) TARGET VFIN IF (NOT *-1 VFIN BARRIER S-BOUNDARY) (0 AUX-OR-MAIN  OR MOD-ASP);
	## Maid Máret áiggui goarrumis?
	## Mun galggan ruoktot.


There was a problem:
"galget" did not receive a syntactic tag in a very clear position in the sentence:

BáikkálaÅ¡ bargit ja hálddahus galget maid sáhttit oahpásmuvvat persovnnalaÅ¡ fierpmádagaiguin, mat doibmet sámiid veahkkevuogádahkan.

The reason was that a very general rule had been removed on 07.11 (version 1.864):

MAP (@+FAUXV) TARGET VFIN IF (0 AUX);
	## Don Å¡attat vel joksat Sirpmás gohcciid.

And replaced by the following rule:

MAP (@+FAUXV) TARGET AUX IF (NOT *-1 AUX BARRIER NOT-AA)(0 VFIN)(*1 Inf OR ConNeg OR PrfPrc BARRIER S-BOUNDARY OR VFIN);
    ## Don Å¡attat vel joksat Sirpmás gohcciid.
	## In sáhte olles gáhku borrat.


"maid" was causing a problem in such an early rule, since it can be both be an adverb and a pronoun. Used in a barrier, it causes a problem. Probably, we will get the same problem with other adverbs that can be other POS. We will have to add them to the same set and disambiguate later, or design rules only for those and disambiguate later.

That is how the set looks like now:


SET NOT-AA = WORD - A - Adv | ("maid") ;