Made a new rule: >SELECT Indef IF (-1 Neg)(0 ("goabbáge")); 	## Ii goabbáge njuolggadus čujut man láhkavuoÄuin daid leat cealkán.
Made a rule in cycle 1:
>REMOVE VGen IF (0 Der/Pass OR Der/goahti OR Der/st OR Der/l);
The problem gets in sentences like: "Eanagirjjit mat čállojuvvo jagiin ..." ere it should be "čállojuvvojit". Now the verb gets ConNeg, which isnÂ´t true. This text contents so many languageerrors that we perhaps remove it from the corpus. Added also other Der to the rule.

Added (NOT 0 VGen) becauseof Muosat:
>REMOVE Prop IF (0 CNOUN)(0 (Prop Pl))(NOT 0 VGen);
  	## Sápmelaččaide lea hui mávssolaÅ¡ ahte sin dárbbut váldojit vuhtii.

#Â áibbas
>SELECT ("áibbas") IF (0 ("áibat"));
	## Maiddái áibbas oÄÄaáigásaÅ¡ ealáhusvuogit.


Made rule for lex.homonomy:
>SELECT ("álo") IF (0 ("állu"));

We have to discuss what to do with these. Now I have made removerules because that they are semantically not quite the same:
>REMOVE ("geavvat") IF (0 ("geavahit"));
>REMOVE ("ássat") IF (0 ("ásahit"));
>REMOVE ("gohččut") IF (0 ("gohčodit"));
>REMOVE ("guorrat") IF (0 ("guorahallat"));