Made a new rule: >SELECT Indef IF (-1 Neg)(0 ("goabbÃ¡ge")); 	## Ii goabbÃ¡ge njuolggadus Äujut man lÃ¡hkavuoÄuin daid leat cealkÃ¡n.
Made a rule in cycle 1:
>REMOVE VGen IF (0 Der/Pass OR Der/goahti OR Der/st OR Der/l);
The problem gets in sentences like: "Eanagirjjit mat ÄÃ¡llojuvvo jagiin ..." ere it should be "ÄÃ¡llojuvvojit". Now the verb gets ConNeg, which isnÂ´t true. This text contents so many languageerrors that we perhaps remove it from the corpus. Added also other Der to the rule.

Added (NOT 0 VGen) becauseof Muosat:
>REMOVE Prop IF (0 CNOUN)(0 (Prop Pl))(NOT 0 VGen);
  	## SÃ¡pmelaÄÄaide lea hui mÃ¡vssolaÅ¡ ahte sin dÃ¡rbbut vÃ¡ldojit vuhtii.

#Â Ã¡ibbas
>SELECT ("Ã¡ibbas") IF (0 ("Ã¡ibat"));
	## MaiddÃ¡i Ã¡ibbas oÄÄaÃ¡igÃ¡saÅ¡ ealÃ¡husvuogit.


Made rule for lex.homonomy:
>SELECT ("Ã¡lo") IF (0 ("Ã¡llu"));

We have to discuss what to do with these. Now I have made removerules because that they are semantically not quite the same:
>REMOVE ("geavvat") IF (0 ("geavahit"));
>REMOVE ("Ã¡ssat") IF (0 ("Ã¡sahit"));
>REMOVE ("gohÄÄut") IF (0 ("gohÄodit"));
>REMOVE ("guorrat") IF (0 ("guorahallat"));


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@10083 c7155fb1-f0a7-4240-a2fc-2600b6f42f90