Removed Der/Pass from V-DER set. Made set V-ACT as V - Der/Pass.
STV renamed to STVLIST, and STV defined as STVLIST - Der/Pass. The idea here is that STV shall be strict transitive, and we did not want the passive versions of the strict transitives to be counted as transitive.

Changed the lex-not-der rules for verbs:

<  REMOVE V-DER IF (0 LEX-V)(NOT 0 (eaddji) OR Der/Pass);
>  REMOVE V-DER IF (0 LEX-V)(NOT 0 (eaddji));
>  REMOVE V-ACT IF (0 Pass);

These changes are tentative, we'll return.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@10200 c7155fb1-f0a7-4240-a2fc-2600b6f42f90