Made new set > LIST DAT-PRON = ("dat") ("dát") ("diet") ("duot") ("dot") ; and referred to this instead of ("dat")m as earlier, in a large number of rules.
A Attr, not Adv, inside an NP:
> Remove Adv IF (-1 Gen)(0 (A Attr))(1 N);

Made a VERY general rule for Dem vs. Pron:
> SELECT Dem IF (0 (Pron Pers Sg3 Gen @GN>));

The trick is to go for Dem whenever @GN>. It is admittedly general, but with our present
corpus we have no counter examples.


Relaxed the VFIN condition to V for oktii:
< REMOVE V IF (*-1 VFIN BARRIER S-BOUNDARY OR CC)(0 ("oktii"));
> REMOVE V IF (*-1 V BARRIER S-BOUNDARY OR CC)(0 ("oktii"));


Rules for Dem if modifying N Rel sequences. This is a robuster rule than the previous one.
> REMOVE Pers IF (0 (Dem Gen))(1 N)(*2 Rel BARRIER NOT-COMMA);

The same goes for this one: Go for a secure right context.
> REMOVE Pers IF (0 Dem LINK *1 N BARRIER NPNH LINK *1 (CS @CS-VP));

NP internal agreement:
< SELECT (Dem Acc) IF (0 ("dat"))(1C (Num Acc) OR (Indef Acc));
> SELECT (Dem Acc) IF (0 DAT-PRON)(1C (Num Acc) OR (Indef Acc));
> SELECT (Dem Acc) IF (0C (@DN>))(1C Acc);


Actio Ess not in Gen contexts:
> REMOVE Ess IF (NOT *-1 VFIN)(0 (Actio Gen))(*1 N BARRIER NPNH);

Generalised the Po rule even more, now also contrasting against Adv.
< SELECT Po IF (-1 (@GP>))(0 Pr)(NOT *1C (@GP) BARRIER NPNH);
> SELECT Po IF (-1 (@GP>))(0 Pr OR Adv)(NOT *1C (@GP) BARRIER NPNH);


Made a rule for bureaucratic prose. It goes for Pl1 rather than Sg2.
We need a sme-tdis.rle BEFORE the sme-dis.rle, looking at the preceeding sentence.
> REMOVE (Prs Sg2) IF (0 (Prs Pl1));

Case agreement NP-internally:
> SELECT (Dem Gen @DN>) IF (*1C (N Ill) BARRIER NOT-ADJ);
> SELECT (Dem Gen @DN>) IF (*1C (N Sg Loc) BARRIER NOT-ADJ);
> SELECT (Indef Gen @PronN>) IF (*1C (N Gen) BARRIER NOT-ADJ);
> SELECT (Indef Sg Nom @PronN>) IF (*1C (N Sg Nom) BARRIER NOT-ADJ);
> SELECT (Indef Pl Nom @PronN>) IF (*1C (N Pl Nom) BARRIER NOT-ADJ);
> SELECT (Indef Gen @PronN>) IF (*1C (N Ill) BARRIER NOT-ADJ);
> SELECT (Indef Gen @PronN>) IF (*1C (N Sg Loc) BARRIER NOT-ADJ);