Several changes: Fixed the grade labels Gq, G2, G3, they had been accidently "upgraded" to G7, G8, G9 recently. Cosmetics, only.
Changed the tt:t rule for Ruotta so that it really hits only Ruotta, and not e.g. "váttis".
A textbook example of the difference between phonology and morphophonology.

Added rule for rl:rll. "Gradation: Cluster r + l"

Commented in again "Illative Singular Metaphony2".
This could be a bad idea, it was a dead-end, but now remains open. Check this, it
hings on the behavoiour of iija and stahta illatives.

Added test bed for Sijte.