Added words to sets.
Made a new rule for choosing "ovcci" instead of "okci":
>SELECT Num IF (*-1 ("diibmu") BARRIER S-BOUNDARY)(0 ("okci"));

Made a new rule:
>SELECT Gen IF (0 ("dii"))(1 Num);
	## Mun boran dii. 7.

Wanted VFIN instead or Actor to "gÃ¡rvodan" in this sentence: "Mun lihkan dii. 7, gÃ¡rvodan ja boran iÄitbiepmu." There were some work to do:

Added (NOT 0 VFIN) to several SELECT case-rules.

Changed from NOT *-2 to NOT *-1 in the first parantese - I will check in the corpus how it works:
>REMOVE VFIN IF (NOT *-1 Rel OR CS OR COMMA BARRIER S-BOUNDARY2)(*-1C VFIN BARRIER S-BOUNDARY2 OR PUNCT LINK NOT *-1 Interr BARRIER NOT-NP)(*1 S-BOUNDARY2 OR CC OR EOS BARRIER NP-HEAD-NOM OR NP-HEAD-ACC);
	## Muhto sÃ¡Äuid son boaldÃ¡ ÄÃ¡skatmeahttun dolas.

Added (OR COMMA) to this rule:
>SELECT V-SG1 IF (*-1 CC OR COMMA BARRIER VERB LINK *-1C V-SG1 BARRIER VERB)(NOT 0 Der/Pass);
	## Das maÅÅÃ¡ dajan juoidÃ¡ Ã¡lbmotrievtti mearkkaÅ¡umi birra, ja geavahan das veahkkin oÄÄaseamos riektegÃ¡ldu.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@12723 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: a240b7b3ccb11cd40ba6d07f414f1a9ffbba58b1