In Hki, working with Ilona on twol matters. Going back to the good old times, or, to be precise, to 1.138. The offending rules "Postvocalic j Surfacing" "Word Final Consonant Neutralization 1" "Deletion: Final h s j Å¾ in Contracted Stems" all got their versions from 1.138 back, pasted in on top of each respective rule set.
This is thus NOT a mechanical revert to 1.138. All other changes since
1.138 are kept; I just reintroduced the 1.138 rule for this concrete
case.

In addition, two test cases are corrected, as follows:
< !â¬ almmÃ¡jY5
> !â¬ albmÃ¡jY5
< !â¬ olmmoÅ¡Y5
> !â¬ olbmoÅ¾Y5
... we had accidently added the wrong base forms.

TODO: look at whatever was the reason for going from 138 to 139, and
take it from there. The present version is the best one there is, at
the moment.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@12895 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: f17cc05fe48b4fe306a3198744d26ea79dae51d8