Worked with the disambiguation of "nuorra".
Added "navdit" to the OPREDV-set.

Made a new set:
>LIST ESS-PROB-ADVL-N = "nuorra" "boaris" ;
# These nouns in essiv can get @ADVL.

Made a new mappingrule:
>MAP (@ADVL)	TARGET Ess IF (0 ESS-PROB-ADVL-N)(NOT *-1 OPREDV BARRIER SV-BOUNDARY)(NOT *1 OPREDV BARRIER SV-BOUNDARY);

Made new rules:
	# nuorra (vs. nuorrat V)
>REMOVE V IF (0 ("nuorra") LINK *-1 V-NOT-COP BARRIER SV-BOUNDARY2);
>REMOVE V IF (0 ("nuorra") LINK *1 V-NOT-COP BARRIER SV-BOUNDARY2);
	## Son boÄii dohko nuorran.

Made some exceptions in this rule son it wonÂ´t hit "nuorat" Pl Nom, but IÂ´m not sure about this rule at all, and I will check in the Victorio-corpus how it works.
>REMOVE Nom IF (0 Comp)(*1 HNOUN BARRIER NPNH OR PUNCT LINK 0 NOT-NOM LINK NOT 0 Inf)(NOT 1 COMMA OR CC)(NOT 0 (Pl Nom) LINK *-1 (V Pl3) BARRIER NPNHA)(NOT 0 (Pl Nom) LINK *1 (V Pl3) BARRIER NPNHA)(NOT 0 (Sg Nom) LINK *-1 V-SG3 BARRIER NPNHA)(NOT 1 Gen LINK *1 Po OR PREGEN BARRIER NP-BOUNDARY)(NOT 1 Dem);
	## Ãlggahit eará doaibmabijuid gáibida baicca lassi árjja...

Made a new rule:
>SELECT (Pl Nom) IF (0 Comp LINK *1 (V Pl3) BARRIER VFIN OR S-BOUNDARY2);
	## Lea buoret borramuÅ¡ go daid maid nuorat dán áigge njillet.

Removed a rule. It is not very good because it hits i.g.: Subj + lea + A + ja + A + placeADVL. The example sentence works without it.
#REMOVE Nom IF (*-1 Nom BARRIER S-BOUNDARY2 OR NPNHA LINK NOT 0 Dem)(0 (A Attr))(1 N);
	## Sápmelaččat leat oadjebasat olggut birasáitagiid ektui.