Worked getting the right analyzis of "maÅÅil", "vuoru", "boaÄus" ja "0-0" in
MaÅÅil lÃ¡ivves vuosttaÅ¡ vuoru boaÄus lei  0-0. and
MaÅÅil lÃ¡ivves vuosttaÅ¡ vuoru lei boaÄus  0-0.

Added (NOT -1 "boaÄus") - removed NOT (GQ<) because it doesnÂ´t function in a mapping rule:
< MAP (@NumN<) TARGET (Num Nom) IF (-1 N OR ACR LINK NOT 0 (@GQ<))(0 CARDINAL
S)(1 EOS)(NOT -1 NUMMAR OR MANNU);
---
> MAP (@NumN<) TARGET (Num Nom) IF (-1 N OR ACR LINK NOT 0 ("boaÄus"))(0 CARD
INALS)(1 EOS)(NOT -1 NUMMAR OR MANNU);

< MAP (@NumN<) TARGET (Num Nom) IF (-1 N OR A OR ACR LINK NOT 0 (@GQ<))(0 CAR
DINALS)(NOT 0 ("oktii"))(NOT 1 NPROP OR A OR ACR)(NOT -1 NUMMAR OR MANNU);
---
> MAP (@NumN<) TARGET (Num Nom) IF (-1 N OR A OR ACR LINK NOT 0 ("boaÄus"))(0
 CARDINALS)(NOT 0 ("oktii"))(NOT 1 NPROP OR A OR ACR)(NOT -1 NUMMAR OR MANNU)

Added (NOT 1 Nom) to:
< MAP (@GP<) TARGET Gen IF (*-1 Pr BARRIER NOT-NP OR CLB LINK NOT -1 Gen)(NOT
 1 (N Gen))(NOT 0 Pron) (NOT 0 TIME LINK *-1 Po BARRIER NP-BOUNDARY LINK -1 G
en);
---
> MAP (@GP<) TARGET Gen IF (*-1 Pr BARRIER NOT-NP OR CLB LINK NOT -1 Gen)(NOT
 1 (N Gen) LINK NOT 0 Nom)(NOT 0 Pron) (NOT 0 TIME LINK *-1 Po BARRIER NP-BOU
NDARY LINK -1 Gen);

Added (NOT *-1 Pr BARRIER NPNH) to this:
< MAP (@GN>) TARGET Gen IF (NOT -1 NUMERALS LINK 0 Num) (NOT 0 ALU OR Num OR
Cmpnd  OR ("ovttas#rÃ¡ÄÄi"))(*1 N BARRIER NOT-AA OR CLB)(NOT 1 Inf);
---
> MAP (@GN>) TARGET Gen IF (NOT -1 NUMERALS LINK 0 Num)(NOT *-1 Pr BARRIER NP
NH)(NOT 0 ALU OR Num OR Cmpnd  OR ("ovttas#rÃ¡ÄÄi"))(*1 N BARRIER NOT-AA OR CL
B)(NOT 1 Inf);
4447c4447

Made rules for boahtu vs. boaÄus
> SELECT ("boaÄus") IF (-1 COPULAS)(0 ("boahtu"))(1 Num);
> SELECT ("boaÄus") IF (0 ("boahtu"))(1 COPULAS)(2 Num);
>       ## MaÅÅil lÃ¡ivves vuosttaÅ¡ vuoru lei boaÄus 0-0.

Changed this rule:
< REMOVE Pr IF (NOT 1 Gen OR Date);
---
> REMOVE Pr IF (NOT *1 Gen OR Date BARRIER NPNH);

Removed this rule - it doenÂ´t seem logic to me:
< REMOVE Gen IF (*-1C Ord BARRIER NPNH OR Gen OR CC OR PUNCT OR REALCLB OR Or
d LINK NOT 0 ("okta"))(NOT 0 VERB OR ("nr") OR ("nummar") OR ("s") OR TIME)(N
OT 1C Po)(NOT *1 N BARRIER NPNH);
---
> #REMOVE Gen IF (*-1C Ord BARRIER NPNH OR Gen OR CC OR PUNCT OR REALCLB OR O
rd LINK NOT 0 ("okta"))(NOT 0 VERB OR ("nr") OR ("nummar") OR ("s") OR TIME)(
NOT 1C Po)(NOT *1 N BARRIER NPNH);


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@13462 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: 4c431fb36b3852a107263ac1db5a1f9b9fda3d41