The task is to give @N< to Com in sentences like: Moai Birehiin barge mánáidgárddis. Ãndde-guovttos Rihtáin gávnnadeigga gávpogis. Báhpa-guovttis eamidiin boÄiiga árrat.
Made a new set:
>LIST DUPRON-NOTPX = (Pron Du1) (Pron Du2) (Pron Du3) ;
And a new rule:
>MAP (@N<) TARGET (N Sg Com) IF (-1 DUPRON-NOTPX OR ("Ãndde-#guovttos") OR ("Ãndde-#guovttis"))(0 HUMAN-INDIVIDUAL)(1 VFIN);

But "Ãndde-#guovttos" and "Ãndde-#guovttis" should be replaced in the rule with som kind of tag. Could Saara help?

Made one mapping rule into two rules:
#MAP (@N> @P<) TARGET (Pron Attr) ;
>MAP (@N> @P<) TARGET (Pron Attr) IF (*-1 Pr) ;
>MAP (@N>) TARGET (Pron Attr) ;

Added Indef to this rule:
>MAP (@N<) TARGET ("eatnat" Indef) IF (-1 ("giittos"));
	## Giittos eatnat luoikamis.

Made a new rule:
>SELECT Indef IF (-1 ("giittos"))(0 ("eatnat"));

"Bures boahtin Äáhppes Bearralii, Miss Turner!"
Made a new rule:
>SELECT Adv IF (0 ("bures") LINK 1 ("boahtin"));
	# Instead of Interj.

and made a constraint about it in this rule:
>REMOVE Adv IF (NOT *-1 VERB BARRIER S-BOUNDARY2)(NOT 0 ("gosa") OR ("mielde"))(NOT 0 ("bures") LINK 1 ("boahtin"))(NOT 1C A)(NOT *1 VERB BARRIER S-BOUNDARY2);