Beacause of expressions like "1600-logu birrasiid" I added "NOT 0 ("biras" Pl Loc)" to this rule: >REMOVE Gen IF (0 N LINK NOT 0 ABSTR-PLACE OR ACTIVITY OR TOPIC OR HUMAN OR TOPIC OR PROCESS OR INDUSTRY OR INSTITUTION OR ORGANIZATION OR TIME)(1 GENERAL-PLACE OR POLITICAL-PLACE LINK NOT 0 ("biras" Pl Loc));
Added OR NP-HEAD-ACC to this rule. I changed BOC to BOS. We have to be aware of that BOC excludes the Pron Rel, and thats not good in many rules. We should check of all the rules with BOC.
>SELECT Acc IF (*-1 BOS BARRIER NPNHA OR Nom OR NP-HEAD-ACC)(*1 V-SG3 BARRIER NOT-NOM)(*2 TV BARRIER (@OBJ) LINK *1 EOS OR S-BOUNDARY2 BARRIER (@OBJ)) ;

Made a new rule for
>SELECT (@+FAUXV) IF (0 ("leat"))(*1 PrfPrc OR (Actio Ess) BARRIER SV-BOUNDARY2);
	## SÃ¡pmelaÄÄat ledje orron dÃ¡in guovlluin 1600-logu birrasiin.

Have discovered that the analyze of PrfPrc doesnÂ´t work - if it is @-FMAINV or @N>. IÂ´ll start working with it.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@13673 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: 0f29b9fe2f8efcbeda5c1de6683341ca3d2f66ed