Now we have 3000 rules! I donÂ´t know if we should be happy or cry...
Worked with disambuation of maÄe/maÄi and daÄe/daÄi which now are added to the subjunction-lexicon, according to Nickel and Nielsen.

I made new sets:
> LIST MADE = "maÄe" "maÄi" ;
> LIST DADE = "daÄe" "daÄi" ;

and added DADE to the CRD- and CRD2-sets.

I also added DADE to this mappingrule
> MAP (@CNP @CVP) TARGET CS IF (0 ("go") OR ("vai") OR XGO OR DADE);

In sentences with MADE and DADE, I give @CVP to both of them, even if there is no VFIN on both sides, because it is an ellipse and the @CVP makes the disambuation of the other members of the sentence easier. This solution is for discussion!

According to this, I made rules:

SELECT CS IF (0 MADE)(*1 DADE) ;
	## MaÄe mielde oaivi, daÄe mielde ságat.
REMOVE CS IF (0 MADE)(NOT *1 DADE) ;

> SELECT CS IF (*-1 MADE BARRIER S-BOUNDARY)(0 DADE);
>       ## MaÄe stuorit RAM-muitu du dihtoris lea, daÄe eanet dihtorprográmmaid sáhtát geavahit oktanaga.

> REMOVE CS IF (NOT *-1 MADE BARRIER S-BOUNDARY)(0 DADE);

> SELECT Adv IF (-1 Dem OR N OR A OR Pron LINK 0 Gen)(0 ("maÄe"))(NOT *1 DADE BARRIER S-BOUNDARY);
>       ## Mun bohten dan maÄe árrat.

> REMOVE LCRD IF (0 DADE) ;

"maÄÄi" is a noun:
> REMOVE N IF (0 MADE)(1 (A Comp))(*2 DADE);
>       ## MaÄe nuorat váhnemat, daÄe deaÅ§aleabbo lea mánáid sámegielmáhttu.

Added (NOT *-1 MADE BARRIER S-BOUNDARY) to this rule:
> SELECT Adv IF (*-1 COPULAS BARRIER NOT-ADV-PCLE LINK NOT *-1 MADE BARRIER S-BOUNDARY)(1 A);
>       ## Sámi riekteáddejupmi ii lean doarvái miellagiddevaÅ¡ suokkarduvvot ovttas dainna gažaldagain.

Added S-BOUNDARY2 to the first barrier in this rule:
> REMOVE Acc IF (*-1C V BARRIER VFIN OR Inf OR S-BOUNDARY2 LINK NOT 0 (PrfPrc @N>) OR V-TRANS-ACT OR NEGFOC)(NOT 0 TIME)(NOT *1 V-TRANS-ACT BARRIER SV-BOUNDARY2);